requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
python-dotenv==1.0.0
alembic==1.12.1
requests==2.31.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4